[Perpetuals] fix maintenance mode, extract GsnSwitch
